+++ src/lib/features/segment/createSegmentService.ts
@@ -1,9 +1,9 @@
 import { Db, IUnleashConfig } from '../../server-impl';
 import { EventService, SegmentService } from '../../services';
 import FakeEventStore from '../../../test/fixtures/fake-event-store';
-import { ISegmentService } from '../../segments/segment-service-interface';
+import { ISegmentService } from './segment-service-interface';
 import FeatureStrategiesStore from '../feature-toggle/feature-toggle-strategies-store';
-import SegmentStore from '../../db/segment-store';
+import SegmentStore from './segment-store';
 import FakeSegmentStore from '../../../test/fixtures/fake-segment-store';
 import FakeFeatureStrategiesStore from '../feature-toggle/fakes/fake-feature-strategies-store';
 import {
