+++ src/lib/features/segment/segment-service.ts
@@ -1,32 +1,32 @@
-import { IUnleashConfig } from '../types/option';
+import { IUnleashConfig } from '../../types/option';
 import {
     IFlagResolver,
     IUnleashStores,
     SKIP_CHANGE_REQUEST,
     SYSTEM_USER,
-} from '../types';
-import { Logger } from '../logger';
-import NameExistsError from '../error/name-exists-error';
-import { ISegmentStore } from '../types/stores/segment-store';
-import { ISegment } from '../types/model';
-import { segmentSchema } from './segment-schema';
+} from '../../types';
+import { Logger } from '../../logger';
+import NameExistsError from '../../error/name-exists-error';
+import { ISegmentStore } from './segment-store-type';
+import { ISegment } from '../../types/model';
+import { segmentSchema } from '../../services/segment-schema';
 import {
     SEGMENT_CREATED,
     SEGMENT_DELETED,
     SEGMENT_UPDATED,
-} from '../types/events';
-import User from '../types/user';
-import { IFeatureStrategiesStore } from '../features/feature-toggle/types/feature-toggle-strategies-store-type';
-import BadDataError from '../error/bad-data-error';
+} from '../../types/events';
+import User from '../../types/user';
+import { IFeatureStrategiesStore } from '../feature-toggle/types/feature-toggle-strategies-store-type';
+import BadDataError from '../../error/bad-data-error';
 import {
     ISegmentService,
     StrategiesUsingSegment,
-} from '../segments/segment-service-interface';
-import { PermissionError } from '../error';
-import { IChangeRequestAccessReadModel } from '../features/change-request-access-service/change-request-access-read-model';
-import { IPrivateProjectChecker } from '../features/private-project/privateProjectCheckerType';
-import EventService from '../features/events/event-service';
-import { IChangeRequestSegmentUsageReadModel } from '../features/change-request-segment-usage-service/change-request-segment-usage-read-model';
+} from './segment-service-interface';
+import { PermissionError } from '../../error';
+import { IChangeRequestAccessReadModel } from '../change-request-access-service/change-request-access-read-model';
+import { IPrivateProjectChecker } from '../private-project/privateProjectCheckerType';
+import EventService from '../events/event-service';
+import { IChangeRequestSegmentUsageReadModel } from '../change-request-segment-usage-service/change-request-segment-usage-read-model';
 
 export class SegmentService implements ISegmentService {
     private logger: Logger;
