+++ src/lib/features/segment/segment-store.test.ts
@@ -1,8 +1,8 @@
-import { ISegmentStore } from '../types/stores/segment-store';
-import dbInit, { ITestDb } from '../../test/e2e/helpers/database-init';
-import getLogger from '../../test/fixtures/no-logger';
-import NotFoundError from '../error/notfound-error';
-import { IUnleashStores, IUser } from '../types';
+import { ISegmentStore } from './segment-store-type';
+import dbInit, { ITestDb } from '../../../test/e2e/helpers/database-init';
+import getLogger from '../../../test/fixtures/no-logger';
+import NotFoundError from '../../error/notfound-error';
+import { IUnleashStores, IUser } from '../../types';
 
 let stores: IUnleashStores;
 let db: ITestDb;
